requirements.txt
python-telegram-bot=21.9
langchain=0.3.13
python-dotenv=1.0.1
langchain-community=0.3.13
arxiv=2.1.3
langchain-mistralai